Move copy button into own folder and extract useCopy hook
